bugfix, ook kleine waarden voor toevoeging of basis behandelen als float

+++ public/class-kleistad-public-recept-beheer.php
@@ -239,7 +239,7 @@ class Kleistad_Public_Recept_Beheer extends Kleistad_ShortcodeForm {
 		$data['recept']['content']['basis']       = [];
 		$basis_limiet                             = count( $basis['basis_component'] );
 		for ( $i = 0; $i < $basis_limiet; $i++ ) {
-			if ( ( '' !== $basis['basis_component'][ $i ] ) && ( 0 !== intval( $basis['basis_gewicht'][ $i ] ) ) ) {
+			if ( ( '' !== $basis['basis_component'][ $i ] ) && ( 0 !== floatval( $basis['basis_gewicht'][ $i ] ) ) ) {
 				$data['recept']['content']['basis'][ $i ] = [
 					'component' => $basis['basis_component'][ $i ],
 					'gewicht'   => str_replace( ',', '.', $basis['basis_gewicht'][ $i ] ) * 1.0,
@@ -249,7 +249,7 @@ class Kleistad_Public_Recept_Beheer extends Kleistad_ShortcodeForm {
 		$data['recept']['content']['toevoeging'] = [];
 		$toevoeging_limiet                       = count( $toevoeging['toevoeging_component'] );
 		for ( $i = 0; $i < $toevoeging_limiet; $i++ ) {
-			if ( '' !== $toevoeging['toevoeging_component'][ $i ] && 0 !== intval( $toevoeging['toevoeging_gewicht'][ $i ] ) ) {
+			if ( '' !== $toevoeging['toevoeging_component'][ $i ] && 0 !== floatval( $toevoeging['toevoeging_gewicht'][ $i ] ) ) {
 				$data['recept']['content']['toevoeging'][ $i ] = [
 					'component' => $toevoeging['toevoeging_component'][ $i ],
 					'gewicht'   => str_replace( ',', '.', $toevoeging['toevoeging_gewicht'][ $i ] ) * 1.0,
